requirements.txt
ajpy==0.0.4
altgraph==0.10.2
appdirs==1.4.3
asn1crypto==0.24.0
backports-abc==0.5
backports.functools-lru-cache==1.6.1
bcrypt==3.1.7
bdist-mpkg==0.5.0
beautifulsoup4==4.9.1
bonjour-py==0.3
bs4==0.0.1
certifi==2019.11.28
cffi==1.12.2
chardet==3.0.4
Click==7.0
colorama==0.4.3
colorlog==4.1.0
configparser==4.0.2
contextlib2==0.6.0.post1
cryptography==2.6.1
dis3==0.1.3
distlib==0.3.0
dnspython==1.16.0
docopt==0.6.2
dropbox==10.2.0
enum34==1.1.6
filelock==3.0.12
Flask==1.1.1
future==0.17.1
futures==3.3.0
get==2019.4.13
getmac==0.8.2
idna==2.6
importlib-metadata==1.5.0
importlib-resources==1.2.0
ipaddress==1.0.22
iptools==0.7.0
itsdangerous==1.1.0
Jinja2==2.11.1
macholib==1.5.1
Mako==1.1.1
MarkupSafe==1.1.1
matplotlib==1.3.1
modulegraph==0.10.4
netifaces==0.10.9
nose==1.3.7
numpy==1.8.0rc1
paramiko==2.7.1
pathlib2==2.3.5
pipenv==2018.11.26
pipreqs==0.4.10
plot==0.6.4
post==2019.4.13
prettytable==0.7.2
public==2019.4.13
py2app==0.7.3
pycparser==2.19
pycryptodome==3.9.7
PyDispatcher==2.0.5
pygame==1.9.6
PyInstaller==3.6
PyNaCl==1.3.0
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==17.2.0
pyparsing==2.0.1
PySimpleGUI==4.19.0
PySocks==1.7.1
python-dateutil==1.5
pytz==2013.7
PyYAML==5.3
query-string==2019.4.13
request==2019.4.13
requests==2.18.4
scandir==1.10.0
scipy==0.13.0b1
singledispatch==3.4.0.3
six==1.12.0
soupsieve==1.9.6
tornado==5.1.1
typing==3.7.4.1
urllib3==1.22
virtualenv==20.0.8
virtualenv-clone==0.5.3
Werkzeug==1.0.0
wget==3.2
xattr==0.6.4
yarg==0.1.9
zipp==1.2.0
zlib-wrapper==0.1.3
